get random game feature added to gamepage
