+++ client/src/components/GetRandom/GetRandom.scss
@@ -1,0 +1,16 @@
+.fa-dice {
+  font-size: 3.7rem;
+}
+
+.dice-container {
+  margin-left: 4rem;
+  flex-grow: 1;
+  display: flex;
+  justify-content: flex-end;
+  color: rgb(180, 180, 180);
+  @media (max-width: 768px) {
+    margin: 0;
+    margin-top: 2rem;
+    justify-content: center;
+  }
+}
