+++ client/src/scenes/GamePage/index.js
@@ -3,9 +3,10 @@ import { useEffect } from "react";
 import axios from "axios";
 import GameDetails from "./components/GameDetails/index";
 import ReactHtmlParser from "react-html-parser";
+import GetRandom from "../../components/GetRandom";
 import "./GamePage.scss";
 
-function GamePage({ game, setGame }) {
+function GamePage({ game, setGame, handleGetRandom }) {
   const { id } = useParams();
 
   useEffect(() => {
@@ -20,20 +21,28 @@ function GamePage({ game, setGame }) {
 
   if (game) {
     return (
-      <div className="game-container">
-        <h1>{game.name}</h1>
-        <div className="top-section">
-          <img src={game.images.medium} alt={`${game.name}`}></img>
-          <GameDetails game={game} />
+      <>
+        <GetRandom handleGetRandom={handleGetRandom}></GetRandom>
+        <div className="game-container">
+          <h1>{game.name}</h1>
+          <div className="top-section">
+            <img src={game.images.medium} alt={`${game.name}`}></img>
+            <GameDetails game={game} />
+          </div>
+          <div className="game-description">
+            <h2>Description</h2>
+            {ReactHtmlParser(game.description)}
+          </div>
         </div>
-        <div className="game-description">
-          <h2>Description</h2>
-          {ReactHtmlParser(game.description)}
-        </div>
-      </div>
+      </>
     );
   } else {
-    return <>loading</>;
+    return (
+      <>
+        <GetRandom handleGetRandom={handleGetRandom}></GetRandom>
+        <>loading</>;
+      </>
+    );
   }
 }
