+++ client/src/scenes/Home/index.js
@@ -1,22 +1,7 @@
 import GameList from "./components/GameList";
 import Form from "./components/Form";
-import { useHistory } from "react-router-dom";
-import axios from "axios";
-import returnId from "../../storedData/randomGames";
 
-function Home({ games, setGames, PopularGames }) {
-  let history = useHistory();
-
-  const handleGetRandom = () => {
-    let id = returnId();
-    axios
-      .get(`${process.env.REACT_APP_SERVER_URL}/getGame`, {
-        params: { id: id },
-      })
-      .then((game) => {
-        history.push(`/game/${game.data.id}`);
-      });
-  };
+function Home({ games, setGames, PopularGames, handleGetRandom }) {
   return (
     <>
       <Form handleGetRandom={handleGetRandom} setGames={setGames}></Form>
